+++ example/package.json
@@ -17,7 +17,7 @@
     "@babel/runtime": "^7.12.5",
     "metro-react-native-babel-preset": "^0.73.5",
     "mkdirp": "^1.0.0",
-    "react-native-test-app": "^2.0.3"
+    "react-native-test-app": "^2.3.1"
   },
   "jest": {
     "preset": "react-native"
