+++ patches/@react-native-segmented-control+segmented-control+2.4.0.patch
@@ -1,0 +1,45 @@
+diff --git a/node_modules/@react-native-segmented-control/segmented-control/js/SegmentedControl.js b/node_modules/@react-native-segmented-control/segmented-control/js/SegmentedControl.js
+index d82ecb5..7555b6a 100644
+--- a/node_modules/@react-native-segmented-control/segmented-control/js/SegmentedControl.js
++++ b/node_modules/@react-native-segmented-control/segmented-control/js/SegmentedControl.js
+@@ -92,6 +92,19 @@ const SegmentedControl = ({
+           selectedIndex={selectedIndex}
+         />
+       )}
++      {selectedIndex != null && segmentWidth ? (
++        <Animated.View
++          style={[
++            styles.slider,
++            {
++              transform: [{translateX: animation}],
++              width: segmentWidth - 4,
++              backgroundColor:
++                tintColor || (colorScheme === 'dark' ? '#636366' : 'white'),
++            },
++          ]}
++        />
++      ) : null}
+       <View style={styles.segmentsContainer}>
+         {values &&
+           values.map((value, index) => {
+@@ -113,19 +126,7 @@ const SegmentedControl = ({
+             );
+           })}
+       </View>
+-      {selectedIndex != null && segmentWidth ? (
+-        <Animated.View
+-          style={[
+-            styles.slider,
+-            {
+-              transform: [{translateX: animation}],
+-              width: segmentWidth - 4,
+-              backgroundColor:
+-                tintColor || (colorScheme === 'dark' ? '#636366' : 'white'),
+-            },
+-          ]}
+-        />
+-      ) : null}
++
+     </View>
+   );
+ };
