+++ src/picker.ios.js
@@ -8,6 +8,6 @@
  * @flow strict-local
  */
 // $FlowFixMe
-import RNDateTimePicker from './specs/DateTimePickerNativeComponent';
+import RNDateTimePicker from './specs/NativeComponentDateTimePicker';
 
 export default RNDateTimePicker;
